+++ src/lib/elastic.js
@@ -4,6 +4,45 @@ const fs = require('fs');
 const jsonFile = require('jsonfile')
 const es = require('elasticsearch')
 
+function _updateQueryStringParameter (uri, key, value) {
+  var re = new RegExp('([?&])' + key + '=.*?(&|#|$)', 'i');
+  if (uri.match(re)) {
+    if (value) {
+      return uri.replace(re, '$1' + key + '=' + value + '$2');
+    } else {
+      return uri.replace(re, '$1' + '$2');
+    }
+  } else {
+    var hash = '';
+    if (uri.indexOf('#') !== -1) {
+      hash = uri.replace(/.*#/, '#');
+      uri = uri.replace(/#.*/, '');
+    }
+    var separator = uri.indexOf('?') !== -1 ? '&' : '?';
+    return uri + separator + key + '=' + value + hash;
+  }
+}
+
+function adjustBackendProxyUrl (req, indexName, entityType, config) {
+  let url
+  if (parseInt(config.elasticsearch.apiVersion) < 6) { // legacy for ES 5
+    url = config.elasticsearch.host + ':' + config.elasticsearch.port + (req.query.request ? _updateQueryStringParameter(req.url, 'request', null) : req.url)
+  } else {
+    const queryString = require('query-string');
+    const parsedQuery = queryString.parseUrl(req.url).query
+    parsedQuery._source_includes = parsedQuery._source_include
+    parsedQuery._source_excludes = parsedQuery._source_exclude
+    delete parsedQuery._source_exclude
+    delete parsedQuery._source_include
+    delete parsedQuery.request
+    url = config.elasticsearch.host + ':' + config.elasticsearch.port + '/' + `${indexName}_${entityType}` + '/_search?' + queryString.stringify(parsedQuery)
+  }
+  if (!url.startsWith('http')) {
+    url = config.elasticsearch.protocol + '://' + url
+  }
+  return url
+}
+
 function adjustQuery (esQuery, entityType, config) {
   if (parseInt(config.elasticsearch.apiVersion) < 6) {
     esQuery.type = entityType
@@ -164,6 +203,7 @@ module.exports = {
   reIndex,
   search,
   adjustQuery,
+  adjustBackendProxyUrl,
   getClient,
   getHits
 }
