+++ src/platform/magento2/tax.js
@@ -49,7 +49,6 @@ class TaxProxy extends AbstractTaxProxy {
     this._taxRegion = taxRegion
     this._sourcePriceInclTax = sourcePriceInclTax
     this._finalPriceInclTax = finalPriceInclTax
-    console.debug('Taxes will be calculated for', taxCountry, taxRegion, sourcePriceInclTax)
     this.taxFor = this.taxFor.bind(this)
   }
